envoyconfig: cleanup
