+++ config/envoyconfig/listeners_envoy_admin.go
@@ -12,10 +12,7 @@ import (
 )
 
 func (b *Builder) buildEnvoyAdminListener(_ context.Context, cfg *config.Config) (*envoy_config_listener_v3.Listener, error) {
-	filter, err := b.buildEnvoyAdminHTTPConnectionManagerFilter()
-	if err != nil {
-		return nil, err
-	}
+	filter := b.buildEnvoyAdminHTTPConnectionManagerFilter()
 
 	filterChain := &envoy_config_listener_v3.FilterChain{
 		Filters: []*envoy_config_listener_v3.Filter{
@@ -28,14 +25,14 @@ func (b *Builder) buildEnvoyAdminListener(_ context.Context, cfg *config.Config)
 		return nil, fmt.Errorf("envoy_admin_addr %s: %w", cfg.Options.EnvoyAdminAddress, err)
 	}
 
-	li := newEnvoyListener("envoy-admin")
+	li := newListener("envoy-admin")
 	li.Address = addr
 	li.FilterChains = []*envoy_config_listener_v3.FilterChain{filterChain}
 	return li, nil
 }
 
-func (b *Builder) buildEnvoyAdminHTTPConnectionManagerFilter() (*envoy_config_listener_v3.Filter, error) {
-	rc, err := b.buildRouteConfiguration("envoy-admin", []*envoy_config_route_v3.VirtualHost{{
+func (b *Builder) buildEnvoyAdminHTTPConnectionManagerFilter() *envoy_config_listener_v3.Filter {
+	rc := newRouteConfiguration("envoy-admin", []*envoy_config_route_v3.VirtualHost{{
 		Name:    "envoy-admin",
 		Domains: []string{"*"},
 		Routes: []*envoy_config_route_v3.Route{
@@ -54,9 +51,6 @@ func (b *Builder) buildEnvoyAdminHTTPConnectionManagerFilter() (*envoy_config_li
 			},
 		},
 	}})
-	if err != nil {
-		return nil, err
-	}
 
 	return HTTPConnectionManagerFilter(&envoy_http_connection_manager.HttpConnectionManager{
 		CodecType:  envoy_http_connection_manager.HttpConnectionManager_AUTO,
@@ -67,5 +61,9 @@ func (b *Builder) buildEnvoyAdminHTTPConnectionManagerFilter() (*envoy_config_li
 		HttpFilters: []*envoy_http_connection_manager.HttpFilter{
 			HTTPRouterFilter(),
 		},
-	}), nil
+	})
+}
+
+func shouldStartEnvoyAdminListener(options *config.Options) bool {
+	return options.EnvoyAdminAddress != ""
 }
